[REFACTOR] MemberUpdate Scene View 리팩토링
- MemberUpdate Scene View 리팩토링
